+++ config.php
@@ -10,9 +10,9 @@ return [
     'urlManagerRules' => [
         ['class' => 'humhub\modules\wiki\components\WikiPageUrlRule'],
     ],
-    'events' => array(
+    'events' => [
         ['class' => Menu::class, 'event' => Menu::EVENT_INIT, 'callback' => ['humhub\modules\wiki\Events', 'onSpaceMenuInit']],
         ['class' => ProfileMenu::class, 'event' => ProfileMenu::EVENT_INIT, 'callback' => ['humhub\modules\wiki\Events', 'onProfileMenuInit']],
         ['class' => 'humhub\modules\rest\Module', 'event' => 'restApiAddRules', 'callback' => ['humhub\modules\wiki\Events', 'onRestApiAddRules']],
-    ),
+    ],
 ];