+++ docs/CHANGELOG.md
@@ -1,10 +1,18 @@
 Changelog
 =========
 
-1.7.1  (Unreleased)
+
+1.9.0-beta.1 (November 7, 2022)
+-------------------------------
+- Enh #244: Implemented new HumHub v1.13 Content Form Support
+
+
+1.8.0  (November 7, 2022)
 -------------------------
+- Enh #241: Remove deprecated checkbox "regular" style
 - Fix #216: Hide space name on print version
 - Enh #236: Allow update content topics
+- Enh #244: Menu for create a Wiki Page from wall stream
 
 1.7.0  (January 27, 2022)
 -------------------------